backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-environ
django-cors-headers
